add context menus to nft thumbnails
