+++ src/entries/popup/pages/home/NFTs/NFTs.tsx
@@ -35,6 +35,7 @@ import { ROUTES } from '~/entries/popup/urls';
 
 import ExternalImage from '../../../components/ExternalImage/ExternalImage';
 
+import NFTContextMenu from './NFTContextMenu';
 import { fadeOutMask } from './NFTs.css';
 
 const NFTS_LIMIT = 2000;
@@ -218,14 +219,16 @@ export function NFTs() {
                               }}
                             >
                               {rowData.map((asset, i) => (
-                                <NftThumbnail
-                                  imageSrc={getUniqueAssetImageThumbnailURL(
-                                    asset,
-                                  )}
-                                  key={i}
-                                  onClick={() => onAssetClick(asset)}
-                                  index={i}
-                                />
+                                <NFTContextMenu key={i} nft={asset}>
+                                  <NftThumbnail
+                                    imageSrc={getUniqueAssetImageThumbnailURL(
+                                      asset,
+                                    )}
+                                    key={i}
+                                    onClick={() => onAssetClick(asset)}
+                                    index={i}
+                                  />
+                                </NFTContextMenu>
                               ))}
                               {rowData.length < 3 &&
                                 isPaginating &&
@@ -489,18 +492,19 @@ function CollectionSection({
             >
               {section.assets.map((asset, i) => {
                 return (
-                  <NftThumbnail
-                    imageSrc={
-                      // we hold off on providing the src field until opened so that
-                      // we don't request images for collections that are never opened
-                      collectionVisible
-                        ? getUniqueAssetImageThumbnailURL(asset)
-                        : undefined
-                    }
-                    key={i}
-                    onClick={() => onAssetClick(asset)}
-                    index={i}
-                  />
+                  <NFTContextMenu key={i} nft={asset}>
+                    <NftThumbnail
+                      imageSrc={
+                        // we hold off on providing the src field until opened so that
+                        // we don't request images for collections that are never opened
+                        collectionVisible
+                          ? getUniqueAssetImageThumbnailURL(asset)
+                          : undefined
+                      }
+                      onClick={() => onAssetClick(asset)}
+                      index={i}
+                    />
+                  </NFTContextMenu>
                 );
               })}
             </Box>
